dash_ui/requirements.txt
dash
dash-bootstrap-components
databricks-sdk
python-dotenv
databricks-sql-connector